% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapFP-package.R
\docType{package}
\name{bootstrapFP-package}
\alias{bootstrapFP-package}
\title{bootstrapFP: Bootstrap Algorithms for Finite Population Inference}
\description{
Perform bootstrap variance estimation of the Horvitz-Thompson total estimator
in finite population sampling with equal or unequal probabilities.
}
\references{
Mashreghi Z.; Haziza D.; Léger C., 2016. A survey of bootstrap methods in 
finite population sampling. Statistics Surveys 10 1-52.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/rhobis/bootstrapFP/issues}
}

}
\author{
\strong{Maintainer}: Roberto Sichera \email{rob.sichera@gmail.com}

}
