% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_antal_tille.R
\name{bs_antal_tille}
\alias{bs_antal_tille}
\title{Antal and Till? (2011) bootstrap for unequal probability sampling}
\usage{
bs_antal_tille(ys, pks, B, smplFUN, approx_method = c("Hajek",
  "DevilleTille"), ...)
}
\arguments{
\item{ys}{values of the variable of interest for the original sample}

\item{pks}{vector of first-order inclusion probabilities for sampled units}

\item{B}{integer scalar, number of bootstrap resamples to draw from the pseudo-population}

\item{smplFUN}{function used to select the sample according to the
original sampling design. The function must accept the vector of inclusion
probabilities as input and return a vector o 0's and 1's (as the functions of
the library sampling).}

\item{approx_method}{method used to approximate Dkk with function \code{Dkk_aprox()}}

\item{...}{added to ignore useless arguments}
}
\value{
a list of two elements, a vector of K average bootstrap totals and
a vector of K variance estimates.
}
\description{
Draw B bootstrap resamples according to Antal and Till? (2011) direct
bootstap method for unequal probability sampling.
Note that this method do not need a double bootstrap.
}
\references{
Antal, E.; Till?, Y., 2011. A Direct Bootstrap Method for Complex Sampling
Designs From a Finite Population. Journal of the American Statistical Association, 106:494, 534-543,
doi: 10.1198/jasa.2011.tm09767

Antal, E.; Till?, Y., 2014. A new resampling method for sampling designs without
replacement: the doubled half bootstrap. Computational Statistics, 29(5), 1345-1363.
doi: 10.10007/s00180-014-0495-0
}
