% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntalTille_UPS.R
\name{decompose_phi}
\alias{decompose_phi}
\title{Decompose phi vector}
\usage{
decompose_phi(phi)
}
\arguments{
\item{phi}{vector of inclusion probabilities for Antal and Till? (2011) bootstrap, 
given by 1 - D_kk}
}
\value{
a list with the two vectors in which \code{phi} is decomposed
}
\description{
Decompose vector phi in a convex combination of two vectors phi1 and phi2, 
such that the sum of $phi1_i$ is the integer part of phi and the sum of $phi2_i$
is the integer part of phi plus 1 (see Antal and Till? (2011) bootstrap procedure
for unequal probability sampling, p. 539 - Algorithm 4, Case 1)
}
\keyword{internal}
