% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_SRS.R
\name{directBS_srs}
\alias{directBS_srs}
\title{Direct bootstrap methods for simple random sampling}
\usage{
directBS_srs(y, N, B, method)
}
\arguments{
\item{y}{vector of sample values}

\item{N}{scalar, represents the population size}

\item{B}{scalar, number of bootstrap replications}

\item{method}{a string indicating the bootstrap method to be used, available
methods are: 'Efron', 'McCarthySnowden', 'RaoWu', 'Sitter'.}
}
\description{
Direct bootstrap methods for simple random sampling
}
\details{
See Mashreghi et al. (2016) for details about the algorithm.
}
\references{
Mashreghi Z.; Haziza D.; Léger C., 2016. A survey of bootstrap methods in 
finite population sampling. Statistics Surveys 10 1-52.
}
