% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntalTille_UPS.R
\name{Dkk_approx}
\alias{Dkk_approx}
\title{Approximation of the Dkk quantity}
\usage{
Dkk_approx(pks, n, method = c("Hajek", "DevilleTille"))
}
\arguments{
\item{pks}{vector of first-order inclusion probabilities for sampled units}

\item{n}{integer, the sample size}

\item{method}{a string indicating which approximation method is to be used}
}
\value{
a numeric vector with approximated Dkk
}
\description{
see Antal and Till? (2011)
}
\keyword{internal}
