% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_antal_tille.R
\name{doubled_half}
\alias{doubled_half}
\title{Select a doubled-half sampling (Antal and Till?, 2014)}
\usage{
doubled_half(n)
}
\arguments{
\item{n}{integer scalar representing sample size}
}
\value{
an integer vector of size \code{n}, indicating how many times each unit is
present in the sample
}
\description{
Select a doubled-half sampling (Antal and Till?, 2014)
}
