% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_hotdeck.R
\name{hd_boot}
\alias{hd_boot}
\title{Hot-deck pseudo-population bootstrap}
\usage{
hd_boot(ys, x, s, B, smplFUN, ...)
}
\arguments{
\item{ys}{values of the variable of interest for the original sample}

\item{x}{population values of the size variable}

\item{s}{logical vector, TRUE for units in the sample, FALSE otherwise}

\item{B}{integer scalar, number of bootstrap resamples to draw from the pseudo-population}

\item{smplFUN}{function used to select the sample according to the
original sampling design. The function must accept the vector of inclusion
probabilities as input and return a vector o 0's and 1's (as the functions of
the library sampling).}

\item{...}{added to ignore useless arguments}

\item{D}{integer scalar, number of replicates for the double bootstrap. 
For original Holmberg's procedure D=1, the default.}
}
\value{
a list of two elements, a vector of K average bootstrap totals and
a vector of K variance estimates.
}
\description{
Draw B bootstrap resamples according to Hot-deck pseudo-population method.
Note that this method do not need a double bootstrap.
}
