% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_wholenumber}
\alias{is_wholenumber}
\title{Check if a number is integer}
\usage{
is_wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{a scalar or a numeric vector}

\item{tol}{a scalar, indicating the tolerance}
}
\description{
Check if \code{x} is an integer number, differently from \code{is.integer},
which checks the type of the object \code{x}
}
\note{
From the help page of function \code{\link[base]{is.integer}}
}
\keyword{internal}
