% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntalTille_UPS.R
\name{over_replacement}
\alias{over_replacement}
\title{Select a simple random sampling with over-replacement}
\usage{
over_replacement(N, n)
}
\arguments{
\item{N}{integer, the population size}

\item{n}{integer, the sample size}
}
\value{
an integer vector of size n, indicating how many times each unit is
present in the sample
}
\description{
Used for resampling procedures. Proposed by Antal and Tille' (2011).
}
\references{
Antal, E.; Tillé, Y. (2011). Simple random sampling with over-replacement. 
Journal of Statistical Planning and Inference, 141(1), 597-601.
}
\keyword{internal}
