% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_population.R
\name{ppBS_srs}
\alias{ppBS_srs}
\title{Pseudo-population bootstrap for simple random sampling}
\usage{
ppBS_srs(y, N, B, D = 1, method)
}
\arguments{
\item{y}{vector of sample values}

\item{N}{scalar, represents the population size}

\item{B}{scalar, number of bootstrap replications}

\item{D}{scalar, number of replications for the double bootstrap (when applicable)}

\item{method}{a string indicating the bootstrap method to be used, available
methods are: 'Gross', 'Booth', 'ChaoLo85', 'ChaoLo94', 'BickelFreedman', 'Sitter'}
}
\description{
Pseudo-population bootstrap for simple random sampling
}
\details{
See Mashreghi et al. (2016) for details about these bootstrap methods.
}
\references{
Mashreghi Z.; Haziza D.; Léger C., 2016. A survey of bootstrap methods in 
finite population sampling. Statistics Surveys 10 1-52.
}
