% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_Uc}
\alias{select_Uc}
\title{Select the random part of a pseudo-population}
\usage{
select_Uc(..., method)
}
\arguments{
\item{...}{parameters of the function, depending on the bootstap method chosen.}

\item{method}{string indicating the bootstrap method}
}
\description{
Helper function that generates the fixed part of a pseudo-population in 
function /code{ppBS_srs()}.
}
\keyword{internal}
